# core/docs/api/openapi.yaml
openapi: 3.0.0
info:
  title: Crawlab API
  version: 0.6.0
  description: Crawlab API
  license:
    name: BSD-3-Clause
    url: https://github.com/crawlab-team/crawlab-pro/blob/main/LICENSE
servers:
  - url: http://localhost:8000
    description: Local API
  - url: http://localhost:8080/api
    description: Docker API
  - url: https://demo-pro.crawlab.cn/api
    description: Demo API
...
paths:
...
  /users/{id}/change-password:
    post:
      tags: [ user ]
      operationId: changePassword
      summary: Change password
      security:
        - apiToken: [ ]
      parameters:
        - name: id
          in: path
          description: User ID
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostUserChangePasswordRequest'
      responses:
        200:
          description: Change password successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
        404:
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundErrorResponse'
        400:
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestErrorResponse'
  /users/me:
    get:
      tags: [ user ]
      operationId: getMe
      summary: Get me
      security:
        - apiToken: [ ]
      responses:
        200:
          description: Get me successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetUserResponse'
        400:
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestErrorResponse'
    put:
      tags: [ user ]
      operationId: putMe
      summary: Update me
      security:
        - apiToken: [ ]
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PutUserRequest'
      responses:
        200:
          description: Update me successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
        400:
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestErrorResponse'

  # token
...
components:
  schemas:
...
    PutUserRequest:
      description: Put user request body
      properties:
        username:
          type: string
          description: User username
        email:
          type: string
          description: User email
        role:
          type: string
          enum:
            - admin
            - user
          description: User role
    PostUserChangePasswordRequest:
      description: Post user change password request body
      properties:
        password:
          type: string
          description: New password
...
    BaseResponse:
      type: object
      properties:
        status:
          type: string
          description: Status of the response
        message:
          type: string
          description: Message of the response
    EmptyResponse:
      description: Empty response body
      $ref: '#/components/schemas/BaseResponse'
      example:
        status: ok
        message: success
    Response:
      description: Response body
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              type: object
...
    ErrorResponse:
      description: Error response body
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            error:
              type: string

    # error response
    NotFoundErrorResponse:
      description: Not found error response body
      $ref: '#/components/schemas/ErrorResponse'
      example:
        status: error
        message: not found
    BadRequestErrorResponse:
      description: Bad request error response
      $ref: '#/components/schemas/ErrorResponse'
      example:
        status: error
        message: bad request
...
    GetUserResponse:
      type: object
      allOf:
        - $ref: '#/components/schemas/Response'
        - properties:
            data:
            $ref: '#/components/schemas/User'
      example:
        {
          "status": "ok",
          "message": "success",
          "data": {
            "_id": "62c665fb6e54352bf1b279e2",
            "username": "admin",
            "role": "admin",
            "email": ""
          },
          "error": ""
        }
...
    Model:
      description: Base model
      type: object
      properties:
        _id:
          type: string
          description: ID of the model (MongoDB ObjectID)
...
    User:
      description: User model
      type: object
      allOf:
        - $ref: '#/components/schemas/Model'
        - properties:
          username:
            description: User username
            type: string
          email:
            description: User email
            type: string
          role:
            description: User role
            type: string
...
  securitySchemes:
    apiToken:
      type: apiKey
      in: header
      name: Authorization
      description: API Token
      x-displayName: API Token